CHROMIUM: mesa: Remove code that returns a desktop GL enum
When the draw_buffer_enum_to_bitmask function
is called while Mesa is in it's meta state,
the wrong GLbitfield is returned sometimes for
OpenGL ES contexts.  This is because the
_mesa_meta_begin function makes the ctx->API always
equal to desktop GL during the meta state.

In the case of GL_BACK, the _mesa_is_gles(ctx)
function is checked when returning the enum.
This check evaluates to false during the meta state,
even on mobile devices.

BUG=b/27656575
TEST=Ran dEQP-GLES2.functional.clipping.line.line_attrib_clip
with the pbuffer option with on Cyan.  Verified it
fails without the patch.

Signed-off-by: Tomasz Figa <tfiga@chromium.org>