RFC: anv: Drop vkCreateDmaBufImageINTEL()
This "extension" function predates Vulkan 1.0 and Vulkan WSI. We created
it during the early days of bringing up the Vulkan driver. We used it
for testing the early driver because, in those early days, there was no
other way to import/export a tiled VkImage and examine it. No
VkSurfaceKHR existed yet in the early API.

We never created a extension for this function.

Since there is no extension, the validation layers don't understand it.
I've observed the validation layers crash when this function is called.

We now support a real dma_buf extension,
VK_MESAX_external_image_dma_buf, that can replace this function.