BACKPORT: FROMLIST: configure.ac/meson.build: Add options for library suffixes
When building the Chrome OS Android container, we need to build copies
of mesa that don't conflict with the Android system-supplied libraries.
This adds options to create suffixed versions of EGL and GLES libraries:

libEGL.so -> libEGL${egl-lib-suffix}.so
libGLESv1_CM.so -> libGLESv1_CM${gles-lib-suffix}.so
libGLESv2.so -> libGLES${gles-lib-suffix}.so

This is similar to what happens when --enable-libglvnd is specified, but
without the side effects of linking against libglvnd.  To avoid
unexpected clashes with the suffixed appended by libglvnd, make it an
error to specify both --enable-libglvnd and --with-egl-lib-suffix.

(am from https://patchwork.freedesktop.org/patch/229368/)
Backport: trivial merge conflicts

BUG=b:78687332
TEST=test_that cheets_CTS_P.9.0_r1.x86.CtsJniTestCases passes

Change-Id: I94a39f6ec177894838a0fcdd050b682250dab564
Signed-off-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Eric Engestrom <eric.engestrom@intel.com>
Signed-off-by: Drew Davenport <ddavenport@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1157883
Reviewed-by: Tomasz Figa <tfiga@chromium.org>