Revert "anv: Use absolute timeouts in wait_for_bo_fences"
This fixes dEQP-VK.synchronization.events on Nami.  Note
this test was removed from dEQP a while ago, and this
revert is needed to satisfy N CTS.

However, it's interesting to note the test still passes on
a v4.13 Intel setup, while it fails on Nami/Fizz (v4.4 devices).
Could be a real bug, but it's pretty easy to revert either
way.

BUG=b:130195490
TEST=./deqp-vk --deqp-case=dEQP-VK.synchronization.events passes on Nami

This reverts commit 1bd4f8fefc2728963fc37900fe75210ee24e09d1.