dri: Limit the max_num_back to 2 on COMPLETE_MODE_FLIP present mode
This is limiting the number of back buffers that mesa can allocate, so
this avoids triple buffering, although that is desirable in some cases.

To get this to upstream, we could convert it to a DRI option
and enable it only in the case of using mutter.
It seems to be feasible to limit this to some kind of configuration, as
we have access to the size of the back-buffer allocated. For example,
only limit for 4k-dual screen setup.

With this Raspberry OS start-up CMA usage is 210Mb with 4k-dual screen
setup instead of 276Mb.

The correct approach would be to check if we can make Mutter to wait
for buffer swaps before starting a new frame.

https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/7033