v3d: Add driconf options to rewrite SCANOUT usages
We create a new eviroment variable V3D_IGNORE_SCANOUT_USAGES
that will affect v3d_resource_create_with_modifiers so
SCANOUT usages can be ignored. It can be enabled under X11
with a compositor so applications are forces to use tiled render
buffers instead of the default behaviour that uses SCANOUT and
consume the limited CMA memory in the RPi4.

The two new driconf options modulate the effect on two applications
Xorg and mutter.

"v3d_maintain_ignorable_scanout": is enabled in mutter, could be used
in other compositors, the objective is that the enviroment has enable
the V3D_IGNORE_SCANOUT_USAGES, they aren't ignored in the compositor.

"v3d_ignore_scanout_usages_with_modifiers": handles a particular case
of the X server when dmabuf_capability is enabled, as the gbm
allocations to back the pixmaps surfaces don't enable the SCANOUT at
Mesa 20.3.5 we just ignore the SHARE usage too so this pixmaps surfaces
are allocated using a tiled format.

Following patches will use this configuration options to ignore or not
the SCANOUT usage on v3d_resource_allocation with modifiers.

Upstreaming this patch need to review the effects of:
 ad50b47a14e9 ("gbm: assume USE_SCANOUT in create_with_modifiers")